move config flag & code into touchbase command
