+++ dev/config/config.go
@@ -16,10 +16,47 @@ google:
     bucket: "kronus"
     prefix: "kronus-dev"
     sqliteBackupSchedule: "*/30 * * * *"
-    enableSqliteBackupAndSync: true
+    enableSqliteBackupAndSync: false
   applicationCredentials:
 
 twilio:
   key:
   number:
+`
+
+const DEFAULT_TOUCHBASE_YML = `
+settings:
+ timezone: "America/Toronto"
+ touchbase-recurrence: "RRULE:FREQ=WEEKLY;"
+
+# Here you update your contact list with their names.
+# e.g.
+# contacts:
+# - name: Smally
+# - name: Dad
+#
+contacts:
+
+# Here you add the different groups you'd like to have for your
+# contacts. And populate each group with 
+# each contact's id(i.e. index of their record in contacts)
+# e.g. 
+# groups:
+#   friends:
+#     - 0
+#     - 1
+#   family:
+#     - 2
+#
+groups:
+
+
+# This section is automatically updated by the CLI App to manage
+# events created by kronus
+events:
+
+owner:
+ email: <The email associated with your google calendar>
+secrets:
+  GOOGLE_APPLICATION_CREDENTIALS: <Path to the JSON file that contains your service account key>
 `
